refactor(mga): extract callout tables into lineup_callouts.py
CI's hard 1000-LOC limit caught reconcile_agent.py at 1002. Allowlisting it
was available — the allowlist explicitly permits constant tables, and 630 of
those lines are exactly that — but the file has a better split than the limit
was asking for.

The tables and four pure string->zone helpers had THREE consumers, and two of
them reached the code by exec()ing reconcile's source with its CLI block
excised by sentinel. That coupled a builder and a test suite to reconcile's
module-level code with nothing in the import graph to show it, and it had
already broken once in this branch: deriving the root from __file__ raised
NameError because exec() supplies no module globals. A real module makes all
three an ordinary import and removes the sentinel dance entirely.

  reconcile_agent.py  1002 -> 375 LOC
  lineup_callouts.py       650 LOC  (7 map tables + 4 pure helpers)

stand_zone_from_loc() deliberately stays in reconcile: it closes over the
CLI-bound CALLOUTS global, so it is not pure and does not belong in a table
module shared with code that runs before the CLI does.

test_callouts.py comes into the repo with them — it is the suite that proves
the tables are right, and it was the third exec() consumer. 823/823 pass
against the extracted module, including the "491/491 table entries are
reachable" and fail-loud-on-unknown-map contracts.

check-file-size now reports all files within thresholds.
